% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electsys21-package.R
\docType{package}
\name{electsys21-package}
\alias{electsys21}
\alias{electsys21-package}
\title{electsys21: Voting Methods for Ranked, Rated and Approval Ballots}
\description{
Implements a range of voting methods and electoral systems for determining election winners, including the D21 Method, First-Past-The-Post, Two-round Runoff, Instant Runoff, the Borda Count, Approval Voting, Majority Judgment and the Condorcet Method. The functions accept several ballot formats - ranking, cardinal utilities, approvals and scores - with automatic detection of the input type, configurable tie-breaking and tidy summaries of the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ivan-ih21/electsys21}
  \item Report bugs at \url{https://github.com/ivan-ih21/electsys21/issues}
}

}
\author{
\strong{Maintainer}: Ivan Iakimov \email{ivan.iakimov@ih21.org}

}
\keyword{internal}
